code/core/maven_junit_jdbc/empapp_jdbc_again/src/main/java/com/empapp/service/EmployeeServiceImpl.java
package com.empapp.service;

import com.empapp.dao.Employee;
import com.empapp.dao.EmployeeDao;
import com.empapp.dao.impl.EmployeeDaoImplJbdc;

import java.util.List;

public class EmployeeServiceImpl implements EmployeeService{

    private EmployeeDao employeeDao;

    public EmployeeServiceImpl(){
        employeeDao=new EmployeeDaoImplJbdc();
    }

    @Override
    public List<Employee> getAll() {
        return employeeDao.getAll();
    }

    @Override
    public void addEmployee(Employee employee) {
        employeeDao.addEmployee(employee);
    }

    @Override
    public void deleteEmployee(int id) {
        employeeDao.deleteEmployee(id);
    }

    @Override
    public void updateEmployee(int id, double salary) {
        employeeDao.updateEmployee(id, salary);
    }

    @Override
    public Employee findById(int id) {
        return employeeDao.findById(id);
    }
}
